[Unit] Restore get promotion test
